Added picture upload and a converter to decode it from db + other minor changes

+++ learningPlatform/src/main/java/com/example/learningPlatform/services/QuestionService.java
@@ -3,17 +3,15 @@ package com.example.learningPlatform.services;
 import com.example.learningPlatform.model.Question;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.IOUtils;
 import org.apache.tomcat.util.codec.binary.Base64;
-import org.hibernate.type.descriptor.jdbc.IntegerJdbcType;
-import org.hibernate.type.descriptor.jdbc.VarcharJdbcType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.*;
 import java.util.ArrayList;
 
 @Service
@@ -27,6 +25,7 @@ public class QuestionService {
     QuestionDataService dataService;
     private ArrayList<Question> questions = new ArrayList<>();
     private Question question = new Question();
+    private boolean loadCheck = true;
 
     public ArrayList<Question> getQuestions(){
         LOG.info(String.valueOf(useData));
@@ -39,34 +38,48 @@ public class QuestionService {
         return questions;
     }
     public Question getQuestion(int id){
-        //System.out.println("Inside getQuestion");
-        Question question1 = dataService.findById(id);
-        //System.out.println(question1);
-        return  question1;
+        LOG.info(String.valueOf(useData));
+        if(useData){
+            question = dataService.findById(id);
+            if(question != null) LOG.info("Got question: " + removeBlobForLogging(question).toString());
+            else LOG.info("Question with id: "+id+" Not found");
+            return question;
+        }
+        return question;
     }
     public String getQuestionBase64String(Question question){
+        LOG.info(String.valueOf(useData));
         String imageString = "";
-        byte[] image = question.getQuestion_photo();
-        if(image.length > 0) imageString = Base64.encodeBase64String(image);
-        return  imageString;
+        if(useData){
+            byte[] image = question.getQuestion_photo();
+            if(image.length > 0){
+                imageString = Base64.encodeBase64String(image);
+                LOG.info("The question has a picture");
+            }else LOG.info("The question does not have a picture");
+            return  imageString;
+        }
+        return imageString;
     }
     public void addQuestion(Question incoming) {
         LOG.info(String.valueOf(useData));
         if (useData) {
             dataService.save(incoming);
-            //Creating a copy of a question and replacing BLOB data, so it does not flood the console (replacing only notEmpty pictures)
-            Question questionSmall = new Question(incoming.getId(),incoming.getTopic(),incoming.getQuestion(),incoming.getAnswers(), incoming.getCorrect_answers(),incoming.getHint(), incoming.getQuestion_photo());
-            if(questionSmall.getQuestion_photo().length > 0) questionSmall.setQuestion_photo("NotEmpty".getBytes());
-            LOG.info(questionSmall.toString());
+            LOG.info("Added question: \n"+removeBlobForLogging(incoming).toString());
             return;
         }
         questions.add(incoming);
     }
     public void deleteQuestionById(int id) {
-        //System.out.println("Inside service deleteQuestionById method");
-        Question test = dataService.findById(id);
-        //System.out.println(test);
-        dataService.deleteById(id);
+        LOG.info(String.valueOf(useData));
+        if (useData) {
+            question = dataService.findById(id);
+            if(question == null){
+                LOG.info("No such question with id: "+id);
+                return;
+            }
+            LOG.info("Deleted question \n: "+removeBlobForLogging(question).toString());
+            dataService.deleteById(id);
+        }
     }
     public String convertToStringLine(String[] answers){
         String line = "";
@@ -79,5 +92,41 @@ public class QuestionService {
             return line;
         }
         return line;
+    }
+    public void populateInitalUploadPhotos() throws IOException {
+        if(loadCheck) { //1st load
+            String[] photoContentRootPath= {"src/main/resources/photos/stop_sign.jpg","src/main/resources/photos/1_GiveWay.png","src/main/resources/photos/2_GiveWay.png"};
+            byte[] filecontent = null;
+            for(int i =0; i< photoContentRootPath.length; i++){
+                File file = new File(photoContentRootPath[i]);
+                FileInputStream inputStream = new FileInputStream(file);
+                filecontent = IOUtils.toByteArray(inputStream);
+                dataService.findById(i+1).setQuestion_photo(filecontent);
+                dataService.save(getQuestion(i+1));
+            }
+            loadCheck=false;
+        }
+    }
+    public byte[] inputImageToBytes(MultipartFile file){
+        LOG.info("File name :" + file.getName() );
+        LOG.info("File size :" + file.getSize() );
+        LOG.info("File type :" + file.getContentType() );
+        byte[] filecontent = null;
+        try {
+            InputStream inputStream = file.getInputStream();
+            if(inputStream == null)
+                LOG.info("file InputStream is null");
+            filecontent = IOUtils.toByteArray(inputStream);
+            return filecontent;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    //Creating a copy of a question and replacing BLOB data, so it does not flood the console (replacing only notEmpty pictures)
+    public Question removeBlobForLogging(Question question){
+        Question questionSmall = new Question(question.getId(),question.getTopic(),question.getQuestion(),question.getAnswers(), question.getCorrect_answers(),question.getHint(), question.getQuestion_photo());
+        if(questionSmall.getQuestion_photo().length > 0) questionSmall.setQuestion_photo("NotEmpty".getBytes());
+        return questionSmall;
     }
 }
